[filmon] Fix harmless -Wformat-security warning
Filmon addon contains a call to snprintf with a format string coming
directly from a function call, hitting "format not a string literal and
no format arguments" warning with -Wformat-security (and thus fails to
build if -Werror=format-security is used as some Linux distributions do).

There does not seem to be a security issue in this case, but modify the
snprintf() call to avoid the warning anyway.